+++ tests/run-almost-all.ts
@@ -113,6 +113,7 @@ import { testit as issue5870ListDirs } from "./issue-5870-list-dirs.ts";
 import { testit as issue5899 } from "./issue-5899.ts";
 import { testit as issue5907 } from "./issue-5907.ts";
 import { testit as issue5934 } from "./issue-5934.ts";
+import { testit as issue5963 } from "./issue-5963.ts";
 import { testit as issue5964 } from "./issue-5964.ts";
 import { testit as issue5965 } from "./issue-5965.ts";
 import { testit as issue5967 } from "./issue-5967.ts";
@@ -193,6 +194,7 @@ export const tests: NamedTest[] = [
   ["issue-5845", issue5845],
   ["issue-5870", issue5870],
   ["issue-5934", issue5934],
+  ["issue-5963", issue5963],
   ["issue-5964", issue5964],
   ["issue-5965", issue5965],
   ["issue-5967", issue5967],